+++ MRML/vtkMRMLROS2SubscriberDefaultNodes.cxx
@@ -6,10 +6,10 @@
 VTK_MRML_ROS_SUBSCRIBER_NATIVE_CXX(std_msgs::msg::String, std::string, String);
 VTK_MRML_ROS_SUBSCRIBER_NATIVE_CXX(std_msgs::msg::Bool, bool, Bool);
 VTK_MRML_ROS_SUBSCRIBER_NATIVE_CXX(std_msgs::msg::Int64, int, Int);
-VTK_MRML_ROS_SUBSCRIBER_NATIVE_CXX(std_msgs::msg::Float64, double, Float);
+VTK_MRML_ROS_SUBSCRIBER_NATIVE_CXX(std_msgs::msg::Float64, double, Double);
 
 VTK_MRML_ROS_SUBSCRIBER_VTK_CXX(std_msgs::msg::Int64MultiArray, vtkIntArray, IntArray);
-VTK_MRML_ROS_SUBSCRIBER_VTK_CXX(std_msgs::msg::Float64MultiArray, vtkDoubleArray, FloatArray);
+VTK_MRML_ROS_SUBSCRIBER_VTK_CXX(std_msgs::msg::Float64MultiArray, vtkDoubleArray, DoubleArray);
 VTK_MRML_ROS_SUBSCRIBER_VTK_CXX(std_msgs::msg::Int64MultiArray, vtkDenseArray<int>, IntNArray);
 
 VTK_MRML_ROS_SUBSCRIBER_VTK_CXX(sensor_msgs::msg::Joy, vtkTable, Joy)
